Rename confirmatio to verifier since it is not a word :-D
