+++ src/components/passphrase/passphraseVerifier.js
