+++ src/components/passphrase/passphraseVerifier.test.js
@@ -3,11 +3,11 @@ import chai, { expect } from 'chai';
 import { spy } from 'sinon';
 import sinonChai from 'sinon-chai';
 import { mount, shallow } from 'enzyme';
-import PassphraseConfirmator from './passphraseConfirmator';
+import PassphraseVerifier from './passphraseVerifier';
 
 chai.use(sinonChai);
 
-describe('PassphraseConfirmator', () => {
+describe('PassphraseVerifier', () => {
   const props = {
     updateAnswer: () => {},
     passphrase: 'survey stereo pool fortune oblige slight gravity goddess mistake sentence anchor pool',
@@ -16,7 +16,7 @@ describe('PassphraseConfirmator', () => {
   describe('componentDidMount', () => {
     it('should call updateAnswer with "false"', () => {
       const spyFn = spy(props, 'updateAnswer');
-      mount(<PassphraseConfirmator passphrase={props.passphrase}
+      mount(<PassphraseVerifier passphrase={props.passphrase}
         updateAnswer={props.updateAnswer} />);
       expect(spyFn).to.have.been.calledWith();
       props.updateAnswer.restore();
@@ -27,7 +27,7 @@ describe('PassphraseConfirmator', () => {
     it('call updateAnswer with received value', () => {
       const spyFn = spy(props, 'updateAnswer');
       const value = 'sample';
-      const wrapper = shallow(<PassphraseConfirmator passphrase={props.passphrase}
+      const wrapper = shallow(<PassphraseVerifier passphrase={props.passphrase}
       updateAnswer={props.updateAnswer}/>);
       wrapper.instance().changeHandler(value);
       expect(spyFn).to.have.been.calledWith();
@@ -37,7 +37,7 @@ describe('PassphraseConfirmator', () => {
 
   describe('hideRandomWord', () => {
     it('should break passphrase, hide a word and store all in state', () => {
-      const wrapper = shallow(<PassphraseConfirmator passphrase={props.passphrase}
+      const wrapper = shallow(<PassphraseVerifier passphrase={props.passphrase}
       updateAnswer={props.updateAnswer}/>);
 
       const randomIndex = 0.5;
